Alexdev

+++ src/app/home/pages/agenda/agenda.component.ts
@@ -1,34 +1,39 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormService } from './services/form.service';
 import { Agenda } from 'src/app/shared/models/agenda.model';
 import { AgendaService } from 'src/app/shared/services/agenda.service';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-agenda',
   templateUrl: './agenda.component.html',
   styleUrls: ['./agenda.component.scss'],
 })
-export class AgendaComponent implements OnInit {
+export class AgendaComponent implements OnInit, OnDestroy {
   agenda: Agenda[] = [];
   selectedItems: Agenda[] = [];
   visible: boolean = false;
+  formSuscription!: Subscription;
 
   constructor(
     private formService: FormService,
     private agendaService: AgendaService,
-    private messageService: MessageService
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService
   ) {}
+  ngOnDestroy(): void {
+    this.formSuscription.unsubscribe();
+  }
   ngOnInit(): void {
     this.loadAgenda();
-    this.formService.open$.subscribe({
+    this.formSuscription = this.formService.open$.subscribe({
       next: (resp) => {
         this.visible = resp.open;
         if (resp.action) {
           console.log(resp.data);
           this.save(resp.data);
         }
-
       },
     });
   }
@@ -60,12 +65,15 @@ export class AgendaComponent implements OnInit {
     if (agendaItem.id) {
       observable = this.agendaService.update(agendaItem).subscribe();
     } else {
-
       observable = this.agendaService.save(agendaItem).subscribe();
-
     }
     observable.add(() => {
       this.loadAgenda();
+      this.formService.onOpen({
+        open: this.visible,
+        action: undefined,
+        data: null,
+      });
       this.messageService.add({
         severity: 'success',
         summary: 'Agenda',
@@ -78,26 +86,35 @@ export class AgendaComponent implements OnInit {
 
   deleteSelected(): void {
     if (this.selectedItems.length === 0) return;
-    this.agendaService.delete(this.selectedItems).subscribe({
-      next: (resp) => {
-        console.log(resp);
-        this.loadAgenda();
-        this.messageService.add({
-          severity: 'success',
-          summary: 'Agenda',
-          detail: 'Actividades eliminadas con éxito!',
+    this.confirmationService.confirm({
+      message: '¿Está seguro que desea eliminar las actividades seleccionadas?',
+      header: 'Confirmación',
+      icon: 'pi pi-exclamation-triangle',
+      acceptLabel: 'Si',
+      rejectLabel: 'No',
+      accept: () => {
+        this.agendaService.delete(this.selectedItems).subscribe({
+          next: (resp) => {
+            console.log(resp);
+            this.loadAgenda();
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Agenda',
+              detail: 'Actividades eliminadas con éxito!',
+            });
+          },
+          error: (err) => {
+            console.log(err);
+            this.messageService.add({
+              severity: 'error',
+              summary: 'Agenda',
+              detail: 'Error al eliminar las actividades!',
+            });
+          },
+          complete: () => {
+            this.selectedItems = [];
+          },
         });
-      },
-      error: (err) => {
-        console.log(err);
-        this.messageService.add({
-          severity: 'error',
-          summary: 'Agenda',
-          detail: 'Error al eliminar las actividades!',
-        });
-      },
-      complete: () => {
-        this.selectedItems = [];
       },
     });
   }
